Remove obsolete utils module
